# refine infershape of sequence_enumerate, hash and fuse_emb_seq_pool